+++ cmd/server/main.go
@@ -12,18 +12,18 @@ import (
 	"github.com/IABTechLab/adscert/internal/logger"
 	"github.com/IABTechLab/adscert/internal/metrics"
 	"github.com/IABTechLab/adscert/internal/utils"
-	"github.com/IABTechLab/adscert/pkg/adscertcrypto"
+	"github.com/IABTechLab/adscert/pkg/adscert/signatory"
 	"github.com/benbjohnson/clock"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"google.golang.org/grpc"
 )
 
 var (
-	serverPort  = flag.Int("server_port", 3000, "grpc server port")
-	metricsPort = flag.Int("metrics_port", 3001, "http metrics port")
-	logLevel    = flag.String("loglevel", utils.GetEnvVar("LOGLEVEL"), "minimum log verbosity")
-	origin      = flag.String("origin", utils.GetEnvVar("ORIGIN"), "ads.cert hostname for the originating party")
-	signatory   adscertcrypto.AuthenticatedConnectionsSignatory
+	serverPort   = flag.Int("server_port", 3000, "grpc server port")
+	metricsPort  = flag.Int("metrics_port", 3001, "http metrics port")
+	logLevel     = flag.String("loglevel", utils.GetEnvVar("LOGLEVEL"), "minimum log verbosity")
+	origin       = flag.String("origin", utils.GetEnvVar("ORIGIN"), "ads.cert hostname for the originating party")
+	signatoryApi signatory.AuthenticatedConnectionsSignatory
 )
 
 func main() {
@@ -32,13 +32,13 @@ func main() {
 	logger.SetLevel(logger.GetLevelFromString(*logLevel))
 
 	// TODO: using randomly generated test certs for now
-	privateKeysBase64 := adscertcrypto.GenerateFakePrivateKeysForTesting(*origin)
+	privateKeysBase64 := signatory.GenerateFakePrivateKeysForTesting(*origin)
 
 	if *origin == "" {
 		logger.Fatalf("Origin hostname is required")
 	}
 
-	signatory = adscertcrypto.NewLocalAuthenticatedConnectionsSignatory(*origin, crypto_rand.Reader, clock.New(), privateKeysBase64, false)
+	signatoryApi = signatory.NewLocalAuthenticatedConnectionsSignatory(*origin, crypto_rand.Reader, clock.New(), privateKeysBase64, false)
 
 	grpcServer := grpc.NewServer()
 	api.RegisterAdsCertServer(grpcServer, &adsCertServer{})
@@ -71,11 +71,11 @@ type adsCertServer struct {
 }
 
 func (s *adsCertServer) SignAuthenticatedConnection(ctx context.Context, req *api.AuthenticatedConnectionSignatureRequest) (*api.AuthenticatedConnectionSignatureResponse, error) {
-	response, err := signatory.EmbossSigningPackage(req)
+	response, err := signatoryApi.EmbossSigningPackage(req)
 	return response, err
 }
 
 func (s *adsCertServer) VerifyAuthenticatedConnection(ctx context.Context, req *api.AuthenticatedConnectionVerificationRequest) (*api.AuthenticatedConnectionVerificationResponse, error) {
-	response, err := signatory.VerifySigningPackage(req)
+	response, err := signatoryApi.VerifySigningPackage(req)
 	return response, err
 }
