+++ pkg/adscert/signatory/crypto_testing.go
@@ -1,4 +1,4 @@
-package adscertcrypto
+package signatory
 
 import (
 	"context"
