+++ pkg/adscert/signatory/local_crypto_impl.go
@@ -1,4 +1,4 @@
-package adscertcrypto
+package signatory
 
 import (
 	"crypto/hmac"
